Try to upgrade to 1.0.0-RC2
