+++ project/build.properties
@@ -1,1 +1,1 @@
-sbt.version=0.13.16-M1
+sbt.version=0.13.16-RC1
