+++ src/main/scala-sbt-1.0/com/typesafe/sbt/packager/MappingsHelper.scala
